refactor(specification): rename module to index
shorter import paths in other modules
