+++ src/syntax/specification/index.ts
