+++ src/syntax/tree/index.spec.ts
@@ -21,7 +21,7 @@ import {
 import { TreeNodeData, TreeNodeStatement, TreeNodeBlock } from './node';
 import { getInstance } from '../warehouse';
 
-import { registerElementSpecificationEntries } from '../specification/specification';
+import { registerElementSpecificationEntries } from '../specification';
 import elementSpecificationEntries from '../../library';
 
 import { ElementProcess, ElementRoutine } from '../../library/elements/elementProgram';
