+++ src/syntax/tree/index.ts
@@ -21,7 +21,7 @@ import {
     TreeNodeBlock,
 } from './node';
 import { addInstance, getInstance, removeInstance } from '../warehouse';
-import { checkValueAssignment, queryElementSpecification } from '../specification/specification';
+import { checkValueAssignment, queryElementSpecification } from '../specification';
 
 import { ElementArgument } from '../elements/elementArgument';
